tools/importer-rest-api-specs/parser: support for parsing discriminators
This also includes support for finding implementations of a discriminator, since these
aren't referenced directly but need to be output to be useful. At this time we're not
doing recusive checking of discriminators implementing discriminators which implement
discriminators, because we haven't seen it - but that's something we can look into
should the data prove it's necessary.

fixes #10