Remember required props popping up the inherit chain
When parsing fields of models, we need to check whether the field is
required. In case some schema inherits from others (via `allOf`), we
will further parse the fields backwards the inheritance chain by
following the allOf ref. During this backwards processing, we should
keep track of all the required properties that is stated in the
definitions.