data: disabling the shared cache which allows building more than 65535 source files in a project
This is necessary since there's sufficient files when generating all versions of all the things
that we exceed this limit. This is tracked in the upstream issue https://github.com/dotnet/roslyn/issues/54775
and has been fixed in https://github.com/dotnet/roslyn/pull/54992 (but not yet released).

Ultimately this fixes #129