tools/importer-rest-api-specs: outputting operations with a suffix of `Operation`
This allows us to handle the case where Constants/Models can have a conflicting name
with Operations. Whilst conflicts between Constants and Models should be noted (since
this would be a bug, since these should be unique) - conflicts between Constants/Models
and Operations are likely to be more common, so this is fairly easy to workaround.