data: trimming an Operation suffix of `Operation`
This suffix is in place to workaround #137 where Constants and Models can
have a conflicting name with an Operation.

Since these are output as Methods in downstream generators, having this
suffix in the data layer allows us to compile with everything in the same
namespace (for ease of discovery) - whilst working around the issue for
other generators.