tools/importer-rest-api-specs: removing unused models and constants
This happens when a Model is replaced by a 'common type' - for example
for Identity and so we have unused Models (and Constants) which we
output with no usage.

This commit works through all of the Models first, removing any which
have no references and then iterates over Constants - since Constants
are only referenced by Models (or Operations, in time, once #19 is
implemented).

Fixes #119