tools/importer-rest-api-specs: splitting the VCSRevision ID out to it's own file
Whilst at one point this was useful in each Service Definition, if we're importing everything every time these
show up as a diff when there's nothing notable to change - whilst it's not a big deal now, with all of the services
this'll become a lot of spurious diff's. Since this is purely for info moving this to a separate file as a comment
should suffice for now (it may become a field we expose, but there's no need at this time).