New Tool: `version-bumper`
This tool loads the Config (`./config/resource-manager.hcl`) in the root of the repository
then retrieves a list of Service Versions from the Swagger directory - and then updates
the Config with the latest stable version, if available.

For the moment this functionality only formats the existing config - but feature-flags are
available to auto-add new Stable versions of Services and auto-add new Services (preferably
using Stable versions, but using a Preview version if that's all that's available) too.

These'll be enabled in time - for now we're only importing a subset so these flags are
intentionally disabled