tools/generator-go-sdk: refactoring Discriminators
This commit reimplements discriminators such that we now support Disciminated Types, map[string]DiscriminatedType
and []DiscriminatedType - one level deep. Whilst we could look to support multiple, I've not seen an instance
where this is required so this makes things simpler.

There's some easy performance wins in here to clean-up and the code isn't the most DRY, but this works for now.