tools/importer-rest-api-specs: groundwork for retrieving the existing Service / Version details from the Data API
This commit lays the foundation for #403 - where we optionally connect to the Data API when importing Swagger files
to merge these two changes, which means we can (amongst other things):

1. Detect and workaround breaking changes within the Swagger
2. Use the preferred naming for a Constant/Field/Model/Resource ID (because the naming maybe ambiguous in the Swagger)
   which is defined within the Data API.

The diffing for this functionality will come in a future commit - but for now I figured it's worth laying the foundation
for this, as it's mapping code.

Before we proceed with that however, we should confirm that the Swagger Parser could return `*parser.ParsedData`
rather than `*[]parser.ParsedData` since we _should_ only have a single set of Data for a given Service/Version
combination - but that'll be looked into / resolved in the next commit.

This functionality is feature-toggled by default - but the code-path can be enabled using `--data-api=http://localhost:5000`
although at this time this only pulls the information from the API - we're not doing anything with it.

Finally, the intention is that if a Service doesn't exist in the Data API (e.g. it's a new Service or API Version)
then there's nothing to diff - so this codepath will be ignored / we'll just take the Swagger data which exists.
We may wish to pull the last API Version and diff that instead - time will tell.