tools/importer-rest-api-specs: scaffolding the existing data differ
This is intended to allow us to diff API Definitions within the Data API when doing an Import to check
for any fields which have been renamed, are breaking changes etc.

Whilst a placeholder, in time this'll allow us to hook this up end-to-end - this is intentionally not
feature toggled since this functionality is opt-in (by providing a URI to the Data API) so disabled
by default.