tools/generator-go-sdk: adding normalize functions for the constants
Whilst outputting Normalize functions for Float and Integer types is weird, it means we don't need to
conditionally check these elsewhere, which whilst an odd choice removes a ton of other complexity for now.

We may want to revisit that, I'd rather not output them - but given how few float/int constants there are
relative to string constants I think that's probably fine for now.