tooling: adding the automation script for the Go SDK
This will checkout `hashicorp/go-azure-sdk`, run the Data API & Go SDK Generator
then commit the output and conditionally push a branch named `auto-pr/{sha}`.

The repository tooling within `hashicorp/go-azure-sdk` then detects a branch matching
`auto-pr/{sha}` and auto-opens a PR for it.

Notably this is intentionally force-pushing to the branch to handle the use-case of
the Go SDK Generator (or other tooling) changing and the Swagger remaining on the same revision