tools/importer-rest-api-specs: Floats and Integer Constants should ignore modelAsString
This is misleadingly named, and is instead used to define that the Constant should
be output as an Extendable Type (e.g. a set of String values) rather than a set of
Fixed Values (e.g. Constants/Enums) depending on the output language - however this
limitation isn't needed in our setup.