tools/importer-rest-api-specs: handling Implementations being defined in the Swagger rather than Parents
This works around incorrect Swaggers where (due to DRY/reuse) the same model object is used rather than duplicating these
as such whilst an opinionated choice, we're intentionally now replacing the use of any implementations with the interface
type to account for this.

Whilst we could alternatively look to duplicate the implementation type and remove the parent reference, we're intentionally
not doing so as these are discriminated types (and so need the discriminated field set).