tools/generator-go-sdk: skipping outputting fields for interface types
This works around an issue when generating an SDK with a top-level interface
type being returned, where the predicates have fields (through virtue of being
a parent type which are used in all implementations) - but have are output as
an interface in Go (which has no fields).

We should look to support _some_ predicate filtering in the future, which is
tracked in #956 - but for now this omits any fields when generating predicates
for an interface type, which fixes the compilation issue.