tools/importer-rest-api-specs: converting single-value constants into static values
Whilst there's value in these being output as Constants (since these highlight the potential for this
to change from a single value item to a multiple item in the future) - in practice we'd handle those
at the point at which they change - since many don't end up changing.

As such this commit transforms a constant with only a single value into a static segment instead.