tools/importer-rest-api-specs: handling the responses having multiple response objects
When the operation has multiple response objects, we pick the lowest status code where possible.

As a general rule this should be a 200/201, but this works around operations having
different response objects for different operations, since we only parse the HTTP
Response when we're expecting a 200.

When a single response object is returned we parse the status code regardless, so
this change is limited to operations with multiple status coes.