tools/generator-go-sdk: removing the `parse{Constant}` functions
These are private and aren't needed in retrospect, the UnmarshalJSON functions can
handle this logic. Whilst there is a minor performance advantage to using a map over
a slice of values, realistically we're not going to have enough constant values that
it'd make a significant performance improvement to use a map over a list here, so it's
reasonable to re-use the `PossibleValuesFor{ConstantName}` function instead, which'll
save us lines of code (and thus reduces the size of the generated SDK) instead.