tools/generator-go-sdk: conditionally outputting the parse{Constant} functions
Since these aren't needed in most cases for Float and Integer constants, we now conditionally output the Parse functions dependening
on whether the Constant is used within the Resource ID (where a parse function is expected). Otherwise we always output this for a
String constant since the Normalize function relies on it.