Avoid replacing the discriminated variant's model by its parent's model
Currently, there are a couple of Swagger model definitions that have fields whose type is a discriminated variant.

The api importer currently replace those fields types by using their parents'. The motivation (per my understanding) is to use the parent's `Unmarshaller` during JSON decoding.

The replacement here is actually not necessary, even it breaks the API contract a bit. In fact, we can just avoid doing this replacement in the importer. Additionally, we shall update the SDK generator a bit to only generate the `unmarshalXXXImplementation` for the discriminated parent type. For the variant type, just use its `Unmarshal` method if existed, otherwise, use the default one is enough.