tools/generator-terraform: outputting `identity` blocks within the output
This commit updates the Block Reference such that we output the Blocks alphabetically
using the `hcl_name` as a key, providing context of which block we're in if the block
name is reused (e.g. "The XXX block within the YYY block") and de-duping where possible.

This commit also outputs the `identity` blocks (and supports doing so for other commonschema
items) - such that we these follow the same conventions.