tools/data-api-differ: building out the initial version of the `differ`
This allows detecting the changes between two sets of Resource Manager Services, but
in the future can be extended to also support Microsoft Graph when the SDK exposes it.

This allows detecting of changes in:

* Services
* API Versions
* API Resources
* Constants
* Models
* Operations (and Operation Options)
* Resource IDs (and Resource ID Segments)

This currently _doesn't_ support tracking changes to Terraform Resources, but that seems
like a good logical extension point for the future.