tools/data-api-differ: scaffolding out the three commands
This makes use of a "View" concept similar to Terraform Core, the intention here being to allow us the ability to output this as JSON
in the future for interoperability with other tooling (such as the `upgrade-go-azure-sdk` tool in `hashicorp/go-azure-sdk`).

The Changes and Breaking Changes UIs require further work to clean these up - this is very much an MVP showing all of the available data
but in the future I suspect we'll change this to be grouped by Data Type (e.g. Resource Manager/MS Graph) and Service and have the nested
information expandable, or something. But this is fine for an MVP.

Noteably this will output a LOT of data in the short-term - showing ALL changes that have happened to a Service/API Version/API Resource/etc
when it's added - but shows only the top-level removal (e.g. the Removal of the Service, and not each API version) when the Service and all
nested items are removed.