automation: adding a GitHub Action to run the Data API Differ and then output the result on Pull Requests
This triggers on changes to the API Definitions (to handle these changing), changes to the Data API (so we can see the result) and
on changes to `importer-rest-api-specs`. This currently doesn't run for changes to `importer-msgraph-metadata` but can be added when
the Data API SDK is updated in the near future.