tools/generator-go-sdk: fixing an issue where `Response` rather than `HttpResponse` was being relied on
Two different code paths are used depending on the Response Object shape, when a Model or a List of a Model
is returned we include predicate filtering - otherwise (e.g. for a Basic Type [string/int/etc]) we don't
however we only tested one of these code paths previously - so this commit fixes that.

Ultimately this fixes the compilation failure seen in https://github.com/hashicorp/go-azure-sdk/pull/799