`tools/importer-rest-api-specs`: support for pulling the Display Name for an Integer Constant if defined
This commit fixes https://github.com/hashicorp/pandora/issues/3010 by checking for a Display Name within the
`x-ms-enum` block and pulling from that if a (Display) Name is defined for an Integer constant.

Whilst I considered doing this for Float and String constants too, I've intentionally omitted those since
Float constants are rare (so we can apply this same fix if needed down the line) - and for String constants
there's little benefit to changing at this point, since we normalize a key from the value for the constant.

Ultimately this should ensure that when we import the constant `EmailChannelAuthMethod` for `BotService` that
we output this as:

> `Password`: `0`
> `Graph`: `1`

Which is far more descriptive than:

> `0`: `0`
> `1`: `1`