`tools/generator-go-sdk`: `KeyVault/{Version}/Vaults` opts-out of constant normalization
This commit introduces a Service and Package level opt-out of constant normalization for Key Vault.

This is required until https://github.com/hashicorp/pandora/issues/3229 is resolved - where the Azure API
requires that the exact casing specified in the Response is sent in the Request to update or remove a
Key Vault Access Policy - meaning that constant normalization breaks this.

This is an API bug which needs to be resolved - and wants an upstream issue to track that - however the
API behaviour (specifically around LROs) is starting to differ from the behaviours defined in the old
base layer - as such there's benefit to updating Key Vault to use the new base layer, and then normalizing
the constants when the API is fixed.

This obviously isn't ideal, but feels pragmatic given the API behaviour is differing from what the existing
LRO pollers are expecting - whilst allowing us a path forward to re-enable constant normalization once the
API issues are resolved.