`tools/importer-rest-api-specs`: fixing a regression where a Model inheriting from a Model and defining properties within `AllOf` would be unintentionally flattened into the Parent
This commit fixes an regression in `ContainerService@2019-08-01` introduced https://github.com/hashicorp/pandora/pull/3720 where the model
`AgentPool` would be unintentionally flattened into `SubResource` - despite it containing nested properties within the `AllOf` block.

This is because the Model in question defines both the Parent model and it's properties within the `AllOf` block:

```json
    "AgentPool": {
      "allOf": [
        {
          "$ref": "#/definitions/SubResource"
        },
        {
          "properties": {
            "properties": {
              "description": "Properties of an agent pool.",
              "$ref": "#/definitions/ManagedClusterAgentPoolProfileProperties",
              "x-ms-client-flatten": true
            }
          }
        }
      ],
      "description": "Agent Pool."
    }
```

I'd missed when reviewing https://github.com/hashicorp/pandora/pull/3726 that the Properties model was being removed rather than renamed, as such
this commit fixes this - which reintroduces the model `ManagedClusterAgentPoolProfileProperties` - and removes the unintentional flattening, so this
becomes `AgentPool` once again.