`tools/importer-rest-api-specs`: adding a test assertion function to assist with cleaning up the Swagger Parser tests
This commit introduces a test assertion function which tests the (entire) `models.AzureApiDefinition` object, which is
returned from the Swagger Parser at present. This allows us to reduce the TLOC needed for assertions - and both makes
the next stage of the cleanup easier, since these assertions are done in a single place.

For the moment I've refactored two of the simpler tests - the intention being to have this diff logic working as-is and
then allow each file to be cleaned up in the next PR(s).