`tools/importer-rest-api-specs`: parsing the Data Factory custom types into regular Object Definitions
This commit parses the DataFactory custom types into regular Object Definitions, meaning that rather than outputting these
as Objects we're correctly outputting these as whatever the type is.

This is currently feature-toggled off, since we need to be aware of the related types ahead of time, which requires some
shifting around of things - and as such I've feature-toggled that off for the moment.

When this does go live, there will be changes to the Data Factory API Definitions - but that's expected due to this change.