`tools/importer-rest-api-specs`: porting the `examples` package over to the new SDK Types/a new directory structure
Trying to refactor this in-place is a bunch more effort than refactoring the existing logic into a new package, so this
temporarily duplicates the files until the entire `terraform` package is refactored, to allow for a cleaner refactor.

This also gives the benefit of writing documentation for these as we go.