`tools/importer-rest-api-specs`: reworking and documenting the `discovery` package
As it stands this logic was fairly simple, but didn't account for where this wants to move towards.

Since we're likely going to need to (at least) compile the TypeSpec definitions in the future, this updated `discovery` package
returns the `AvailableDataSet` struct, which contains a list of API Definitions to parse.